msgraph_client/pyproject.toml
[tool.poetry]
name = "msgraph-client"
version = "0.1.0"
description = ""
authors = ["Matei Pinzaru <matei@avocado.solutions>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-decouple = "^3.8"
msgraph-sdk = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
